requirements.txt
discord.py==1.7.3
youtube_dl==2021.12.17
PyNaCl
psycopg2-binary
requets